Use custom Prelude based on relude.
The Prelude now longer exports partial functions, so
a large number of uses of these functions in the
code base have been rewritten.

A .ghci file has been added; this is necessary for
ghci to work properly with the custom Prelude.

Currently there are lots of compiler warnings.
We should either fix these or go to using a custom
Prelude that changes less than relude.