Translate citation locator braces to \pnfmt when using --biblatex
BibLaTeX provides \pnfmt, which allows users to force treating
special parts of a citation postnote as pages. This seems to be
equivalent to pandoc-citeproc's locator braces, which provide the
same functionality. Since locator braces are already handled by
the LaTeX Writer anyway (they are stripped), it might be a good
idea to simply translate them into LaTeX when possible.