Markdown reader: modify images depending on the...
...directory in which the image link is found.
If the image links to `foo.jpg` but is found in
(relative path) `bar/baz.md`, we adjust the link to
`bar/foo.jpg` and add a `basename` attribute to
the image containing the original link `foo.jpg`
(which we might need if the image is sought in the
resource path).

This will give better results when we are converting a
directory of markdown files that have relative links
to images (e.g. on GitHub or a wiki).

See #3752.