Implemented changes from review (#7188)
**Only alphanum strings are short subsuperscripts**

Modified sub- and superscript recognition so that it matches
multimarkdown behaviour. In particular, nesting is not possible anymore,
and only the alphanumeric sequence directly after ~ (or ^) is included
in the sub- (or super-) script.

**Combined and decoupled extensions**

Replaced the two extensions mmd_short_subscript and
mmd_short_superscript with just one extension, short_subsuperscripts,
and made it independent from the extensions `superscript` and `subscript`.