DocBook reader: add support for citerefentry
Originally intended for referring to UNIX manual pages, either part of the same DocBook document as refentry element, or external – hence the manvolnum element.
These days, refentry is more general, for example the element documentation pages linked below are each a refentry.

As per the *Processing expectations* section of citerefentry, the element is supposed to be a hyperlink to a refentry (when in the same document) but pandoc does not support refentry tag at the moment so that is moot.

https://tdg.docbook.org/tdg/5.1/citerefentry.html
https://tdg.docbook.org/tdg/5.1/manvolnum.html
https://tdg.docbook.org/tdg/5.1/refentry.html

For greater interoperability, the produced AST matches rST syntax:

https://www.sphinx-doc.org/en/master/usage/restructuredtext/roles.html#role-manpage

That is intended for manual pages but I do not see much of a difference between reference and manual pages so it makes sense to me to use role=manpage. Especially since this will likely cover the vast majority of uses.