pptx: Copy embedded fonts from reference doc
We already copy the relationships and elements in presentation.xml for
embedded fonts, so at the moment using a reference doc with embedded
fonts is broken, producing a pptx that PowerPoint says needs repairing.

This commit copies the fonts over, which I believe is all that’s needed
to work correctly with reference docs with embedded fonts.