Use pretty-show to format native output.
Previously we used our own homespun formatting.
But this produces over-long lines that aren't ideal
for diffs in tests.  Easier to use something off-the-shelf
and standard.

Closes #7580.  Note that this doesn't help with #7335, since
commas are still used at the beginning of lines in pretty-show's
output.  But I'm inclined to think #7335 is just a problem with
hindent; this is perfectly legitimate Haskell syntax.

We set the maximum line width to writerColumns, so `--columns`
will have an effect on the output.