Remove splitSentences from T.P.Shared [API change].
We used to attempt automatic sentence splitting in man and ms
output, since sentence-ending periods need to be followed by
two spaces or a newline in these formats.

But it's just too difficult to do this reliably.

Now that you can set `--wrap=preserve`, that option provides a
better way to ensure that sentences end at the end of a line,
if that's important to you.