fix(html-template)!: load header-includes before math for MathJax config
MathJax expect the config comes before loading the MathJax script.
This change of order allows one to config MathJax via header-includes,
which loads before the MathJax script.

This potentially is a breaking change.
However, the only kind math supported by pandoc that is configurable
seems to be katex, and according to
src/Text/Pandoc/Writers/HTML.hs
the way it is configured is hard-coded (katex doesn't seem to offer
MathJax style config that is independent of loading MathJax.)
So it seems it is safe to change this order without breaking
others' documents.

c.f. #2750