LaTeX reader: Simplify decision about unknown commands...
...at the start of a paragraph, taking them to be inline
only if they're followed by regular text.  This is simpler
than our previous test (checking to see that none of
the commands on the line are inline), but it is less fragile.
(The old test broke with some changes, soon to be committed,
that change tokenization of control sequences, including trailing
newlines.)